pitch-main/pyproject.toml
[project]
name = "pitch"
version = "0.1.0"
description = "AI-Driven Pitch Deck Analysis Platform"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-multipart>=0.0.5",
    "aiofiles>=0.7.0",
    "crewai>=0.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "pinecone-client>=2.2.4",
    "openai>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=0.19.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "redis>=4.0.0",
    "aio-pika>=9.0.0",
    "prometheus-client>=0.16.0",
    "python-pptx>=0.6.22",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "websockets>=12.0",
    "python-magic>=0.4.27",
    "clamd>=1.0.2",
    "boto3>=1.26.0",
    "sentry-sdk>=1.0.0",
    "elasticsearch>=8.0.0",
    "loguru>=0.7.0"
]
requires-python = ">=3.8"

[project.scripts]
pitch = "pitch.main:run"
run_crew = "pitch.main:run"
serve = "pitch.main:serve"
train = "pitch.main:train"
replay = "pitch.main:replay"
test = "pitch.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pitch"]
